chore: small reorganizations
