+++ packages/render/src/helpers/escape-string.ts
