+++ packages/render/src/helpers/jsx-to-string.ts
