+++ packages/render/test/fixtures/preview.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 interface PreviewProps {}
 
-export const Preview: React.VFC<Readonly<PreviewProps>> = () => (
+export const Preview: React.FC<Readonly<PreviewProps>> = () => (
   <>
     <div data-id="@jsx-email/preview">This should be hidden from plain text</div>
     <h1>This should be rendered in plain text</h1>
