+++ packages/render/test/fixtures/template.tsx
@@ -4,7 +4,7 @@ interface TemplateProps {
   firstName: string;
 }
 
-export const Template: React.VFC<Readonly<TemplateProps>> = ({ firstName }) => (
+export const Template: React.FC<Readonly<TemplateProps>> = ({ firstName }) => (
   <>
     <h1>Welcome, {firstName}!</h1>
     <img src="img/test.png" alt="test" />
